# Shadow-Ledger-System-main/Shadow-Ledger-Project/Shadow-Ledger-System/api-specs.yaml
openapi: 3.0.0
info:
  title: Shadow Ledger System API
  description: Microservices API for real-time financial transaction tracking and drift correction
  version: 1.0.0
  contact:
    name: HDFC Life Development Team
    email: api-support@hdfclife.com

servers:
  - url: https://api.shadowledger.hdfclife.com
    description: Production server
  - url: https://api-staging.shadowledger.hdfclife.com
    description: Staging server

security:
  - BearerAuth: []

paths:
  # Authentication Endpoints
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "user@hdfclife.com"
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token
                  expiresIn:
                    type: integer
                    description: Token expiration in seconds
                  role:
                    type: string
                    enum: [USER, AUDITOR, ADMIN]
        '401':
          description: Invalid credentials

  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
                - role
              properties:
                username:
                  type: string
                  minLength: 3
                password:
                  type: string
                  minLength: 8
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [USER, AUDITOR, ADMIN]
      responses:
        '201':
          description: User created successfully
        '400':
          description: Invalid input data

  # Event Endpoints
  /events:
    get:
      tags:
        - Events
      summary: List financial events
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

    post:
      tags:
        - Events
      summary: Create a new financial event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid event data

  /events/{eventId}:
    get:
      tags:
        - Events
      summary: Get event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found

  # Shadow Ledger Endpoints
  /accounts/{accountId}/shadow-balance:
    get:
      tags:
        - Shadow Ledger
      summary: Get shadow balance for account
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
        - name: asOfDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Shadow balance information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShadowBalance'

  # Drift Correction Endpoints
  /drift-check:
    get:
      tags:
        - Drift Correction
      summary: Check for drift across accounts
      security:
        - BearerAuth: []
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
        - name: threshold
          in: query
          schema:
            type: number
            format: decimal
            default: 0.01
      responses:
        '200':
          description: Drift analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  drifts:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriftInfo'
                  summary:
                    type: object
                    properties:
                      totalAccounts:
                        type: integer
                      accountsWithDrift:
                        type: integer
                      maxDriftAmount:
                        type: number
                        format: decimal

  /correct/{driftId}:
    post:
      tags:
        - Drift Correction
      summary: Apply drift correction
      security:
        - BearerAuth: []
      parameters:
        - name: driftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                strategy:
                  type: string
                  enum: [BALANCE_ADJUSTMENT, EVENT_REPLAY, MANUAL_CORRECTION]
                reason:
                  type: string
                  maxLength: 500
                approvedBy:
                  type: string
      responses:
        '200':
          description: Correction applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrectionResult'
        '403':
          description: Insufficient privileges

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Event:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        accountId:
          type: string
        transactionType:
          type: string
          enum: [CREDIT, DEBIT, TRANSFER]
        amount:
          type: number
          format: decimal
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: 'USD'
        businessTimestamp:
          type: string
          format: date-time
        systemTimestamp:
          type: string
          format: date-time
        sequenceNumber:
          type: integer
          format: int64
        metadata:
          type: object
          additionalProperties: true

    CreateEventRequest:
      type: object
      required:
        - accountId
        - transactionType
        - amount
        - currency
      properties:
        accountId:
          type: string
        transactionType:
          type: string
          enum: [CREDIT, DEBIT, TRANSFER]
        amount:
          type: number
          format: decimal
          minimum: 0.01
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        businessTimestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true

    ShadowBalance:
      type: object
      properties:
        accountId:
          type: string
        shadowBalance:
          type: number
          format: decimal
        primaryBalance:
          type: number
          format: decimal
        lastUpdated:
          type: string
          format: date-time
        isDriftDetected:
          type: boolean
        driftAmount:
          type: number
          format: decimal

    DriftInfo:
      type: object
      properties:
        driftId:
          type: string
          format: uuid
        accountId:
          type: string
        primaryBalance:
          type: number
          format: decimal
        shadowBalance:
          type: number
          format: decimal
        driftAmount:
          type: number
          format: decimal
        detectedAt:
          type: string
          format: date-time
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]

    CorrectionResult:
      type: object
      properties:
        correctionId:
          type: string
          format: uuid
        driftId:
          type: string
          format: uuid
        strategy:
          type: string
        status:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
        correctedAmount:
          type: number
          format: decimal
        appliedAt:
          type: string
          format: date-time
        appliedBy:
          type: string